Remove 'FK' from all types' names
Because none of the names for types in this framework conflict with the
ones found in the Swift standard library, there isn't much of a reason
to prepend them with 'FK'.

This also makes working with the types feel more at-home with Swift.

Since `FKDataType` has been renamed to `DataType`, the generic type for
`FKFile` (now `File`) has been renamed to `Data` due to conflicts.