Configure Swift Linter
This disables unnecessary rules like `trailing_newline`, also it
disables `valid_docs` for pragmatic reasons: A lot of lines are
documented but their documentation isn't complete. This
option should be reverted sometime but I'm disabling it for the
moment to fix all warnings and come to a clean slate.

On the other hand this enables the rules `empty_count` and
`missing_docs`. This helps even improve documentation
although the `valid_docs` is turned off.