Fix all Swift Linter warnings + Improve docs
This commit fixes many warnings Swift Linter shows
when `valid_docs` is turned on – but not all which is
why it's kept in disabled mode. Also this refactors
the FileSystemEvent file by splitting it up into three
files. Otherwise the Swift Linter shows a warning
because the files has too many files.